+++ crates/multicalc/tests/suite/ode/mod.rs
@@ -1,2 +1,3 @@
+mod exponential_map;
 mod rk4;
 mod rk45;